requirements.txt
-i https://pypi.org/simple
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
google-auth==1.6.0
idna==2.7
pyasn1-modules==0.2.2
pyasn1==0.4.4
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
